Simplify wp_get_attachment_image_sizes internals
We've not found a compelling reason to pass a set of arguments into this function
in order to build a source size list because any time that it would make sense to
do so, e.g. in a template tag, it would be much easier to simply write the sizes
attribute manually, rather than using this function to do the lifting for you.

Instead, we can pass a width parameter representing the display width of the original
image in order to build our default markup, eliminating the need to recalculate the
width using redundant calls to `image_get_intermediate_size()`.

This also removes the need to filter the arguments used to build the sizes attribute,
so we can move the filter to the returned value, which is more expected.

Updated tests.