Limit image creation during upload to default sizes.
WordPress creates intermedia image files during generation of attachment
metadata, which can sometimes cause uploads to fail if the server can't
generate all of the image sizes.

This limits the generation of intermediate image sizes to only default
sizes and 'post-thumbnail' (if applicable) and adds a scheduled event
for generating additional image sizes.

This adds:

* `wp_make_intermediate_images()` – which generates intermediate sizes
for a specific file.
* `wp_generate_attachment_sizes()` - which generates any missing intermediate
image sizes for a specific attachment.