Update a previous crop of an image if one exists
This adds `Custom_Image_Header::get_previous_crop()` which will find a
header for the current theme that was previously cropped from the same
base image by checking the value of  the `_wp_attachment_parent` value
which is now set when inserting an attachment. If a previous crop exists
the new crop overwrites the previous crop, otherwise a new attachment is
created.