PHPUnit: Fix test errors in PHPUnit 6.4.0+
`Tests_Image_Functions::test_load_directory()` used `assertNotInternalType()`
to ensure `wp_load_image()` and `wp_get_image_editor()` wouldn't load
a link to a resource. Under the hood, PHPUnit uses `get_resource_type()`
in this assertion, which will produce an error that was previously not
surfaced. This updates the two assertions to test for what we expect those
functions to return instead.